pre run commit
What is left to do:
fix 1 exception error.
check to see if a special win message needs to be output in a certain way and implement that if so.
Then run the program to make sure it works.
Then implement testing class.
Then add any extra features such as priority queue and ensuring there is permutations of total weights from the files entered that can lead to 100 weight.